Muutin NFmiMenuItemList luokan käyttämään NFmiPtrList luokan sijasta std::list luokkaa. Samalla laitoin dynaamiset menu-itemit std::unique_ptr luokan hallinnoimiksi. Muutin rajapintoja, jotta kaikki tarvittava voidaan toteuttaq sujuvasti std::list luokalla. Esim. Reset-Next iteraatio muutettiin begin-end tyyppiseksi, jotta saadaan käyttöön range-loopitus. Poistin NFmiMenuItemList(NFmiDrawParamList ) kontruktorin, koska sitä ei ole ikinä toteutettu ja se heitti poikkeuksen.
Nämä muutokset heijastuivat muutamiin muihin luokkiin.

+++ libs/fmi/bizcode/smetdata/source/NFmiMenuItem.cpp
@@ -198,7 +198,7 @@ void NFmiMenuItem::IndexInViewRow(int newValue)
 	itsIndexInViewRow = newValue;
 }
 
-unsigned long NFmiMenuItem::NumberOfSubMenuItems(void)
+size_t NFmiMenuItem::NumberOfSubMenuItems(void)
 {
 	if(itsSubMenu)
 		return itsSubMenu->NumberOfMenuItems();
